Kentsel_haber_gorsellestirme/backend/requirements.txt
# Python bağımlılıkları
fastapi
uvicorn[standard]
pymongo
python-dotenv
requests
beautifulsoup4
sentence-transformers
spacy
python-dateutil
